# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * nopad - ephemeral collaboration API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface Room {
  id: string;
  name: string;
  createdAt: string;
  expiresAt: string;
  fileCount: number;
  wordCount: number;
}

export interface FileRecord {
  id: string;
  roomId: string;
  originalName: string;
  storedName: string;
  size: number;
  mimeType: string;
  uploadedAt: string;
}

export interface TextBlock {
  id: string;
  roomId: string;
  content: string;
  ownerId?: string | null;
  committed: boolean;
  updatedAt: string;
  createdAt: string;
}

export interface CreateBlockRequest {
  ownerId: string;
}

export interface UpdateBlockRequest {
  content: string;
  ownerId: string;
}

export interface DeleteBlockRequest {
  ownerId?: string;
}

export interface CommitBlockRequest {
  ownerId: string;
}