# lib/api-zod/src/generated/types/commitBlockRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * nopad - ephemeral collaboration API
 * OpenAPI spec version: 0.1.0
 */

export interface CommitBlockRequest {
  ownerId: string;
}