# lib/api-zod/src/generated/types/fileRecord.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * nopad - ephemeral collaboration API
 * OpenAPI spec version: 0.1.0
 */

export interface FileRecord {
  id: string;
  roomId: string;
  originalName: string;
  storedName: string;
  size: number;
  mimeType: string;
  uploadedAt: Date;
}