# lib/api-zod/src/generated/types/textBlock.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * nopad - ephemeral collaboration API
 * OpenAPI spec version: 0.1.0
 */

export interface TextBlock {
  id: string;
  roomId: string;
  content: string;
  ownerId?: string | null;
  committed: boolean;
  updatedAt: Date;
  createdAt: Date;
}