# lib/api-zod/src/generated/types/updateBlockRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * nopad - ephemeral collaboration API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateBlockRequest {
  content: string;
  ownerId: string;
}